requirements.txt
pydantic==2.5.2
numpy
openai==1.3.7
python-dotenv
scipy
boto3
